Upgrade sha1 to 0.10
This is actually a single major version release from 0.6. The new
version standardizes on the Digest trait used in other digest crates.
That means its API is now consistent with other crates used by this
package.

One notable difference is the new Digest trait doesn't have an easy way
to format the digest to hex. So we just feed the binary data into
hex::encode().